pyproject.toml
[project]
name = "spec2sphere"
version = "2.0.0"
description = "Spec2Sphere — Horvath Analytics Delivery Factory: AI-Governed SAP Datasphere + SAC Delivery Accelerator"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "atlassian-python-api>=3.41",
    "PyGithub>=2.3",
    "openai>=1.0",
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "jinja2>=3.1",
    "bcrypt>=4.0",
    "itsdangerous>=2.1",
    "python-multipart>=0.0.9",
    "pdfplumber>=0.11",
    "python-docx>=1.1",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    "redis>=5.0",
]

[project.scripts]
spec2sphere = "spec2sphere.cli:cli_main"
sap-doc-agent = "spec2sphere.cli:cli_main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "respx>=0.21",
    "ruff>=0.4",
]
anthropic = ["anthropic>=0.25"]
all = ["anthropic>=0.25", "celery[redis]>=5.3", "redis>=5.0", "alembic>=1.13", "psycopg2-binary>=2.9"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"spec2sphere" = [
    "web/templates/*.html",
    "web/templates/**/*.html",
    "web/static/*.css",
    "web/static/*.js",
    "migration/templates/*.html",
    "tenant/templates/*.html",
    "tenant/templates/**/*.html",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 120
